requirements.txt
fastapi
uvicorn[standard]

# Database & ORM
sqlalchemy
psycopg2-binary
alembic

# Authent & Security
passlib[bcrypt]
pyjwt
python-dotenv

# Data Validation
pydantic
email-validator

